requirements.txt
fastapi
motor
pydantic[email]
python-decouple
dnspython
uvicorn